[DOP-3640]: Rename to simplify stuff
