+++ cdk-infra/lib/auto-builder-stack.ts
@@ -1,8 +1,8 @@
 import * as cdk from 'aws-cdk-lib';
 import { Construct } from 'constructs';
-import { AutoBuilderApiConstruct } from './constructs/auto-builder-api-construct';
-import { AutoBuilderQueuesConstruct } from './constructs/auto-builder-queues-construct';
-import { AutoBuilderEnvConstruct } from './constructs/auto-builder-env-construct';
+import { AutoBuilderApiConstruct } from './constructs/api/api-construct';
+import { AutoBuilderQueuesConstruct } from './constructs/queue/queues-construct';
+import { AutoBuilderEnvConstruct } from './constructs/api/env-construct';
 
 export class AutoBuilderStack extends cdk.Stack {
   constructor(scope: Construct, id: string, props?: cdk.StackProps) {
