+++ cdk-infra/lib/constructs/worker/worker-construct.ts
@@ -1,0 +1,7 @@
+import { Construct } from 'constructs';
+
+export class WorkerConstruct extends Construct {
+  constructor(scope: Construct, id: string) {
+    super(scope, id);
+  }
+}
